(PC-30978)[PRO] refactor: Separate styles between table rows and cells.
