+++ pro/src/pages/Offers/OffersTable/Cells/ExpirationCell/ExpirationCell.module.scss
@@ -1,0 +1,3 @@
+.banner {
+  padding: 1rem;
+}
