+++ pro/src/pages/Offers/OffersTable/Cells/IndividualActionsCell.tsx
@@ -1,8 +1,7 @@
 import { ListOffersOfferResponseModel } from 'apiClient/v1'
 import { OFFER_STATUS_DRAFT } from 'core/Offers/constants'
 
-import styles from '../OfferRow.module.scss'
-
+import styles from './Cells.module.scss'
 import { DeleteDraftCell } from './DeleteDraftCell'
 import { EditOfferCell } from './EditOfferCell'
 import { EditStocksCell } from './EditStocksCell'
