+++ pro/src/pages/Offers/OffersTable/IndividualOffersTable/IndividualOfferRow/IndividualOfferRow.module.scss
@@ -1,0 +1,26 @@
+@use "styles/mixins/_rem.scss" as rem;
+@use "styles/variables/_size.scss" as size;
+
+.individual-row {
+  display: grid;
+  grid-template-columns: rem.torem(22px) repeat(3, 1fr);
+  grid-template-areas:
+    "check thumb   head    head"
+    ".     thumb   desc    desc"
+    ".     desc2   desc3   desc3"
+    ".     desc4   desc4   desc4"
+    ".     actions actions status";
+  border-bottom: rem.torem(1px) solid var(--color-grey-medium);
+}
+
+@media (min-width: size.$mobile) {
+  .individual-row {
+    grid-template-columns: rem.torem(22px) rem.torem(112px) 1fr rem.torem(105px);
+  }
+}
+
+@media (min-width: size.$tablet) {
+  .individual-row {
+    display: table-row;
+  }
+}
