+++ pro/src/screens/Offers/Offers.module.scss
@@ -52,7 +52,6 @@
   }
 
   tbody > tr {
-    border-bottom: rem.torem(1px) solid var(--color-grey-medium);
     margin-top: 1rem;
 
     &:first-of-type {
